Fix compile failure on Mac in r11417.
R=jkummerow@chromium.org

Review URL: https://chromiumcodereview.appspot.com/10187003

git-svn-id: http://v8.googlecode.com/svn/branches/bleeding_edge@11419 ce2b1a6d-e550-0410-aec6-3dcde31c8c00

+++ src/isolate.cc
@@ -1857,12 +1857,12 @@ bool Isolate::Init(Deserializer* des) {
     LOG(this, LogCompiledFunctions());
   }
 
-  CHECK_EQ(OFFSET_OF(Isolate, state_),
-           static_cast<intptr_t>(Internals::kIsolateStateOffset));
-  CHECK_EQ(OFFSET_OF(Isolate, embedder_data_),
-           static_cast<intptr_t>(Internals::kIsolateEmbedderDataOffset));
-  CHECK_EQ(OFFSET_OF(Isolate, heap_.roots_),
-           static_cast<intptr_t>(Internals::kIsolateRootsOffset));
+  CHECK_EQ(static_cast<int>(OFFSET_OF(Isolate, state_)),
+           Internals::kIsolateStateOffset);
+  CHECK_EQ(static_cast<int>(OFFSET_OF(Isolate, embedder_data_)),
+           Internals::kIsolateEmbedderDataOffset);
+  CHECK_EQ(static_cast<int>(OFFSET_OF(Isolate, heap_.roots_)),
+           Internals::kIsolateRootsOffset);
 
   state_ = INITIALIZED;
   time_millis_at_init_ = OS::TimeCurrentMillis();
